FM1D2D-951: Refactor RR load mechanism so it used file based system instead of nhibernate. Converted nhibernate tests to filebased working tests Former-commit-id: 7efc37c6ac9bdba7d4547b798320c7e4c1432bd9 Former-commit-id: 43e6380d738c79ffe14854000964dc38eaf93764

+++ src/DeltaShell.Sobek.Readers/Readers/SobekRrReaders/SobekRRIniSettings.cs
@@ -9,7 +9,7 @@ namespace DeltaShell.Sobek.Readers.Readers.SobekRrReaders
         public DateTime EndTime { get; set; }
         public TimeSpan TimestepSize { get; set; }
         public bool PeriodFromEvent { get; set; }
-        public int OutputTimestepMultiplier { get; set; }
+        public double OutputTimestepMultiplier { get; set; }
         public int UnsaturatedZone { get; set; }
         public int InitCapsimOption { get; set; }
         public bool CapsimPerCropAreaIsDefined { get; set; }
